Fix bugs in propagating letter_counts
The first bug is that if the same letter received two Blacks, we need to use
the earlier one as the letter_count endpoint.

The second bug is that the implied letter count from letter choices needs to be
computed to supplement the letter counts. This allows us to deduce more
information.